requirements.txt
beautifulsoup4==4.9.0
boto3==1.10.28
botocore==1.13.28
bs4==0.0.1
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
dj-database-url==0.5.0
Django==2.2.5
django-allauth==0.40.0
django-bootstrap-form==3.4
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-storages==1.8
docutils==0.15.2
gunicorn==19.9.0
idna==2.8
jmespath==0.9.4
oauthlib==3.1.0
Pillow==6.2.0
postgres==3.0.0
psycopg2==2.8.3
psycopg2-binary==2.8.4
psycopg2-pool==1.1
PyJWT==1.7.1
PySocks==1.7.1
python-dateutil==2.8.0
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2019.2
requests==2.22.0
requests-oauthlib==1.2.0
s3transfer==0.2.1
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
soupsieve==2.0
source==1.2.0
sqlparse==0.3.0
tweepy==3.8.0
urllib3==1.25.6
virtualenv==16.7.5
virtualenvwrapper-win==1.2.5
whitenoise==4.1.3
